task2_text_extraction/requirements_task2.txt
paddleocr==2.7.0.3
paddlepaddle==2.6.1
datasets==2.16.1
jiwer==3.0.3
evaluate==0.4.2
python-Levenshtein==0.25.1
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.4.0
tqdm==4.66.5
